chore: clean up page file
